Track cardinality of querysel seal.
This will allow to differentiate between sealed queryset that would benefit
from prefetching many-to-many relationships and the ones that don't.

In other words, warning about N+1 query issues when fetching a many-to-many
relationship from a singular result (.get(), .first()) is arguably bad practice
as it could be beneficial to defer the fetching. This is a common scenario when
seal()'ing a queryset meant to be used both by singular (detail view) and
multiple results (list view) code paths.