Fix #64 -- Adjust SealableReverseOneToOneDescriptor.get_queryset signature.
The override was not matching ReverseOneToOneDescriptor's signature where
instance is not a required parameter and happens not to be passed by
get_prefetch_queryset.

The lack of instance passing by get_prefetch_queryset allows to circumvent
a missing get_object method which would have made a much better injection
point.